+++ app/components/pages/otp/otpRegister/style.scss
@@ -1,92 +1,104 @@
 @import '../../../../style/sass/variable';
 
-$input-width: 39px;
-$input-container-height: 20px;
-$input-height: 18px;
-$handle-width: $input-height;
-$handle-height: $handle-width;
-$primary: $color-toggle;
-$secondary: $color-white;
-$tertiary: $grey-6;
-$border: 1px solid #000;
+// $input-width: 39px;
+// $input-container-height: 20px;
+// $input-height: 18px;
+// $handle-width: $input-height;
+// $handle-height: $handle-width;
+// $primary: $color-toggle;
+// $secondary: $color-white;
+// $tertiary: $grey-6;
+// $border: 1px solid #000;
 
-.chili-animate-field {
-  label {
-    .required-start {
-      color: #929292 !important;
-    }
-  }
-}
+// .chili-animate-field {
+//   label {
+//     .required-start {
+//       color: #929292 !important;
+//     }
+//   }
+// }
 
-.register-otp__pass-box {
-  border: 1px dashed green;
-  border-radius: 14px;
-}
+// .register-otp__pass-box {
+//   border: 1px dashed green;
+//   border-radius: 14px;
+// }
 
-.filters__checkbox {
-  // float: right;
-  vertical-align: middle;
+// .filters__checkbox {
+//   // float: right;
+//   vertical-align: middle;
 
-  input {
-    display: none;
-  }
+//   input {
+//     display: none;
+//   }
 
-  label {
-    position: relative;
-    width: $input-width;
-    right: 10px;
+//   label {
+//     position: relative;
+//     width: $input-width;
+//     right: 10px;
 
-    &::before,
-    &::after {
-      content: "";
-      position: absolute;
-      left: 0;
-      border-radius: $input-height;
-      border: $border;
-      transition: all 0.2s ease;
-      cursor: pointer;
-    }
+//     &::before,
+//     &::after {
+//       content: "";
+//       position: absolute;
+//       left: 0;
+//       border-radius: $input-height;
+//       border: $border;
+//       transition: all 0.2s ease;
+//       cursor: pointer;
+//     }
 
-    &::before {
-      width: $input-width;
-      height: $input-container-height;
-      border: $border;
-      background-color: $tertiary;
-      border-color: $tertiary;
-    }
+//     &::before {
+//       width: $input-width;
+//       height: $input-container-height;
+//       border: $border;
+//       background-color: $tertiary;
+//       border-color: $tertiary;
+//     }
 
-    &::after {
-      content: "";
-      width: $handle-width;
-      height: $handle-height;
-      background: $secondary;
-      border: 0;
-      top: 1px;
-      left: 1px;
-    }
+//     &::after {
+//       content: "";
+//       width: $handle-width;
+//       height: $handle-height;
+//       background: $secondary;
+//       border: 0;
+//       top: 1px;
+//       left: 1px;
+//     }
 
-  }
+//   }
 
-  input:checked+label::after {
-    left: 50%;
-  }
+//   input:checked+label::after {
+//     left: 50%;
+//   }
 
-  input:checked+label::before {
-    background: $primary;
-    border-color: $primary;
-  }
+//   input:checked+label::before {
+//     background: $primary;
+//     border-color: $primary;
+//   }
 
-  &::after {
-    content: attr(data-label);
-    padding-left: 5px;
-    padding-top: 3px;
-  }
+//   &::after {
+//     content: attr(data-label);
+//     padding-left: 5px;
+//     padding-top: 3px;
+//   }
 
-  input:required {
-    background: #000;
-  }
-}
+//   input:required {
+//     background: #000;
+//   }
+// }
 
 .ruleAndCondition{
   color: $grey-1;
+}
+.auth__layout{
+    .header__simple{
+      padding-bottom:0 !important;
+    }
+    &_teacher{
+      .header__simple{
+        > div {
+          background-color: $color-primary !important;
+        }
+      }
+    }
 }
