+++ app/components/snappAuthLayout/index.js
@@ -3,17 +3,17 @@ import PropTypes from 'prop-types';
 
 import './style.scss';
 const SnappAuthLayout = props => {
-  const { className, children, title } = props;
+  const { className, children, title, icon } = props;
   return (
     <div
-      className={`auth__layout whFull authBg bottomP30 ${!!className &&
+      className={`auth__layout wFull authBg bottomP30 ${!!className &&
         className}`}
     >
       <div className="header__simple center rCol bottomP30">
         <div className="icon-bar bg-secondary round50 center">
-          <i className="icon icon-profile text40 color-white" />
+          <i className={`icon text40 color-white ${icon}`} />
         </div>
-        <span className="text18 bold">{title}</span>
+        {!!title && <span className="text18 bold">{title}</span>}
       </div>
       <div className="center rCol">{children}</div>
     </div>
@@ -24,5 +24,9 @@ SnappAuthLayout.propTypes = {
   children: PropTypes.node,
   className: PropTypes.string,
   title: PropTypes.string,
+  icon: PropTypes.string,
+};
+SnappAuthLayout.defaultProps = {
+  icon: 'icon-profile',
 };
 export default SnappAuthLayout;
