+++ app/containers/RegisterPage/RegisterTeacherPage/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import OtpRegister from '../../../components/pages/otp/otpRegister/teacher';
+
+const RegisterIndex = () => <OtpRegister />;
+export default RegisterIndex;
