+++ app/containers/Routers/index.js
@@ -12,6 +12,7 @@ import CoursePage from '../CoursePage';
 import AuthenticationPage from '../AuthenticationPage';
 import VerifyCode from '../VerifyCodePage';
 import RegisterStudent from '../RegisterPage/RegisterStudentPage';
+import RegisterTeacher from '../RegisterPage/RegisterTeacherPage';
 import Register from '../RegisterPage';
 import Profile from '../ProfilePage/mainPage';
 import Page404 from '../Page404';
@@ -28,6 +29,7 @@ function App() {
         <Route exact path="/verify-code" component={VerifyCode} />
         <Route exact path="/register" component={Register} />
         <Route exact path="/register-student" component={RegisterStudent} />
+        <Route exact path="/register-teacher" component={RegisterTeacher} />
         <Route exact path="/profile" component={Profile} />
         <Route component={Page404} />
       </Switch>
